move get_levels
